+++ docs/changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Effort-based Versioning](https://jacobtomlinson.dev
 
 ### Added
 
+- Add option to profile the directory process with [Yappi](https://github.com/sumerc/yappi).
 - Report timings at the end of the worker log files.
 
 ## [3.2.1][] - 2026-01-02 {: #v3.2.1 }